SEOULDATA-65 [FEATURE] 행사 리뷰 등록 구현 (#10)
* feat/SEOULDATA-65 fix: 에러 코드 열거형으로 변환

* feat/SEOULDATA-65 feat: 행사 리뷰 등록 API 구현

* feat/SEOULDATA-65 chore: 테스트 관련 dependency 설정

* feat/SEOULDATA-65 rename: FestRepositoryTest 위치 변경

* feat/SEOULDATA-65 test: ReviewController 테스트 코드 작성

* feat/SEOULDATA-65 rename: 도메인 구조로 클래스 이동 및 분리
