+++ build.gradle
@@ -54,6 +54,7 @@ subprojects {
         testImplementation('org.springframework.boot:spring-boot-starter-test') {
             exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
         }
+        testImplementation("org.assertj:assertj-core:3.22.0")
 
     }
 }
