+++ common/src/main/java/com/seouldata/common/exception/ErrorCode.java
@@ -1,8 +1,18 @@
 package com.seouldata.common.exception;
 
-public interface ErrorCode {
+import lombok.Getter;
 
-    int getCode();
-    String getMessage();
+@Getter
+public enum ErrorCode {
+
+    FEST_NOT_FOUND(400, "축제 정보가 없습니다.");
+
+    private final int code;
+    private final String message;
+
+    ErrorCode(int code, String message) {
+        this.code = code;
+        this.message = message;
+    }
 
 }
