+++ fest/build.gradle
@@ -11,5 +11,4 @@ dependencies {
 	implementation project(":common")
 	implementation 'org.springframework.boot:spring-boot-starter-validation'
 	implementation 'org.springframework.cloud:spring-cloud-starter-openfeign:4.1.0' // openfeign
-// testImplementation(testFixtures(project(':common')))
 }
