+++ fest/src/main/java/com/seouldata/fest/domain/review/service/ReviewService.java
@@ -1,0 +1,9 @@
+package com.seouldata.fest.domain.review.service;
+
+import com.seouldata.fest.domain.review.dto.request.AddReviewReq;
+
+public interface ReviewService {
+
+    Long addReview(Long memSeq, AddReviewReq addReviewReq);
+
+}
